LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/traverseTo-after-data-url-expected.txt



FAIL navigation.traverseTo() should work in an iframe that started at a data: url assert_equals: expected 3 but got 5
